% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{bot_token}
\alias{bot_token}
\title{bot_token}
\usage{
bot_token(botname = NULL)
}
\arguments{
\item{botname}{character of length 1 with the name of the bot}
}
\description{
Obtain token from system variables (in \code{Renviron}) set
according to package naming conventions, that is
\code{R_TELEGRAM_X} where \code{X} is bot's name (first question
answered to the botfather).
}
\examples{
\dontrun{
bot_token('RBot')
}
}

