% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{get_tgbot_token}
\alias{get_tgbot_token}
\title{get_tg_token}
\usage{
get_tgbot_token(botname = NULL)
}
\arguments{
\item{botname}{character of length 1 with the name of the bot}
}
\description{
Obtain token from system variables (in \code{Renviron}) set
according to package naming conventions, that is
\code{R_TELEGRAM_X} where \code{X} is bot's name (first question
answered to the botfather).
}
\examples{
get_tgbot_token('RBot')
}

