% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{getFile}
\alias{getFile}
\title{getFile}
\usage{
getFile(file_id, destfile = NULL)
}
\arguments{
\item{file_id}{File identifier (required)}

\item{destfile}{Destination path; if specified the file will be
downloaded}
}
\description{
Get info about a file and download it
}

