% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_telegram.R
\name{InlineKeyboardButton}
\alias{InlineKeyboardButton}
\title{InlineKeyboardButton}
\usage{
InlineKeyboardButton(text = NULL, url = NULL, callback_data = NULL,
  switch_inline_query = NULL, switch_inline_query_current_chat = NULL)
}
\arguments{
\item{text}{Label text on the button}

\item{url}{HTTP url to be opened when button is pressed}

\item{callback_data}{Data to be sent in a
\href{https://core.telegram.org/bots/api#callbackquery}{callback query}
to the bot when button is pressed, 1-64 bytes}

\item{switch_inline_query}{If set, pressing the button will prompt the
user to select one of their chats, open that chat and insert the
bot's username and the specified inline query in the input field.
Can be empty, in which case just the bot's username will be inserted.}

\item{switch_inline_query_current_chat}{If set, pressing the button will
insert the bot's username and the specified inline query in the current
chat's input field. Can be empty, in which case only the bot's username
will be inserted.}
}
\description{
This object represents one button of an inline keyboard. You
\strong{must} use exactly one of the optional fields.
}
\details{
\strong{Note:} After the user presses a callback button,
Telegram clients will display a progress bar until you call
\code{\link{answerCallbackQuery}}. It is, therefore, necessary to
react by calling \code{\link{answerCallbackQuery}} even if no notification
to the user is needed (e.g., without specifying any of the
optional parameters).
}
