% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_telegram.R
\name{InlineKeyboardMarkup}
\alias{InlineKeyboardMarkup}
\title{InlineKeyboardMarkup}
\usage{
InlineKeyboardMarkup(inline_keyboard = NULL)
}
\arguments{
\item{inline_keyboard}{List of button rows, each represented by a list of
\code{\link{InlineKeyboardButton}} objects}
}
\description{
This object represents an \href{https://core.telegram.org/bots#inline-keyboards-and-on-the-fly-updating}{inline keyboard}
that appears right next to the message it belongs to.
}
\details{
\strong{Note:} After the user presses a callback button,
Telegram clients will display a progress bar until you call
\code{\link{answerCallbackQuery}}. It is, therefore, necessary to
react by calling \code{\link{answerCallbackQuery}} even if no notification
to the user is needed (e.g., without specifying any of the
optional parameters).
}
