% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_telegram_envvar.R
\name{proxy}
\alias{proxy}
\title{proxy}
\usage{
proxy(x)
}
\arguments{
\item{x}{character of length 1 with the name of the proxy; if
\code{NULL} a menu to choose between proxy is displayed and the
proper value returned}
}
\description{
Obtain telegram proxy from system variables (in \code{Renviron}) set
according to package naming conventions, that is
\code{R_TELEGRAM_PROXY_X_Y} where X is the proxy's name and Y are the
proxy parameters.
}
\examples{
\dontrun{ proxy('default') }
}
