% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_telegram.R
\name{ReplyKeyboardMarkup}
\alias{ReplyKeyboardMarkup}
\title{ReplyKeyboardMarkup}
\usage{
ReplyKeyboardMarkup(keyboard = NULL, resize_keyboard = NULL,
  one_time_keyboard = NULL, selective = NULL)
}
\arguments{
\item{keyboard}{List of button rows, each represented by an list of
\code{\link{KeyboardButton}} objects}

\item{resize_keyboard}{Requests clients to resize the keyboard vertically
for optimal fit. Defaults to If \code{FALSE}, in which case the
customkeyboard is always of the same height as the app's standard
keyboard.}

\item{one_time_keyboard}{Requests clients to hide the keyboard as soon
as it's been used. The keyboard will still be available, but clients
will automatically display the usual letter-keyboard in the chat –
the user can press a special button in the input field to see the
custom keyboard again. Defaults to \code{FALSE}}

\item{selective}{Use this parameter if you want to show the keyboard
to specific users only.}
}
\description{
This object represents a \href{https://core.telegram.org/bots#keyboards}{custom keyboard}
with reply options.
}
