% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendAudio}
\alias{sendAudio}
\title{sendAudio}
\usage{
sendAudio(
  audio = NULL,
  caption = NULL,
  duration = NULL,
  performer = NULL,
  title = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)
}
\arguments{
\item{audio}{path to audio file to send (required)}

\item{caption}{audio caption}

\item{duration}{duration of the audio in seconds}

\item{performer}{performer}

\item{title}{track name}

\item{disable_notification}{Sends the message silently. Users will
receive a notification with no sound}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Send \code{mp3} files
}
