% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendChatAction}
\alias{sendChatAction}
\title{sendChatAction}
\usage{
sendChatAction(action = NULL, chat_id = NULL)
}
\arguments{
\item{action}{type of action to broadcast. Choose one, depending on
what the user is about to receive: "typing" for text messages,
"upload_photo" for photos, "record_video" or "upload_video" for
videos, "record_voice" or "upload_voice" for voice notes,
"upload_document" for general files, "find_location" for
location data, "record_video_note" or "upload_video_note" for
video notes.}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Tell the user that something is happening on the bot's side
}
