% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendDice}
\alias{sendDice}
\title{sendDice}
\usage{
sendDice(disable_notification = NULL, chat_id = NULL)
}
\arguments{
\item{disable_notification}{Sends the message silently. Users will
receive a notification with no sound}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Use this method to send a dice, which will have a random value from
1 to 6.
}
