% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendLocation}
\alias{sendLocation}
\title{sendLocation}
\usage{
sendLocation(latitude = NULL, longitude = NULL,
  reply_to_message_id = NULL, chat_id = NULL)
}
\arguments{
\item{latitude}{Latitude of location}

\item{longitude}{Longitude of location}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel}
}
\description{
Send point on the map
}

