% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendMessage}
\alias{sendMessage}
\title{sendMessage}
\usage{
sendMessage(text = NULL, parse_mode = NULL,
  disable_web_page_preview = NULL, reply_to_message_id = NULL,
  reply_markup = NULL, chat_id = NULL)
}
\arguments{
\item{text}{Text of the message to be sent (required)}

\item{parse_mode}{send 'Markdown' if you want Telegram apps to show
bold, italic and inline URLs in your bot's message}

\item{disable_web_page_preview}{Disables link previews for links in
this message}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{reply_markup}{A Keyboard object, it can be ehter:
\itemize{
 \item{\code{\link{ReplyKeyboardMarkup}}}
 \item{\code{\link{InlineKeyboardMarkup}}}
}}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Send text messages
}
