% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendPhoto}
\alias{sendPhoto}
\title{sendPhoto}
\usage{
sendPhoto(photo = NULL, caption = NULL, reply_to_message_id = NULL,
  chat_id = NULL)
}
\arguments{
\item{photo}{photo to send (required)}

\item{caption}{photo caption}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Send image files
}

