% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendPoll}
\alias{sendPoll}
\title{sendPoll}
\usage{
sendPoll(
  question = NULL,
  options = NULL,
  is_anonymous = TRUE,
  type = c("regular", "quiz"),
  allows_multiple_answers = FALSE,
  correct_option_id = NULL,
  is_closed = FALSE,
  chat_id = NULL
)
}
\arguments{
\item{question}{Poll question, 1-255 characters}

\item{options}{vector of possible answers: 2-10 strings 1-100
characters each}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
sendPoll
}
