% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendVideo}
\alias{sendVideo}
\title{sendVideo}
\usage{
sendVideo(video = NULL, duration = NULL, caption = NULL,
  reply_to_message_id = NULL, chat_id = NULL)
}
\arguments{
\item{video}{Video to send (required)}

\item{duration}{Duration of sent video in seconds}

\item{caption}{Video caption}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Send \code{mp4} videos
}

