% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendVoice}
\alias{sendVoice}
\title{sendVoice}
\usage{
sendVoice(voice = NULL, duration = NULL, reply_to_message_id = NULL,
  chat_id = NULL)
}
\arguments{
\item{voice}{Audio file to send (required)}

\item{duration}{Duration of sent audio in seconds}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Send \code{.ogg} files encoded with OPUS
}

