% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{stopPoll}
\alias{stopPoll}
\title{stopPoll}
\usage{
stopPoll(message_id = NULL, chat_id = NULL)
}
\arguments{
\item{message_id}{Identifier of the original message with the poll}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Use this method to stop a poll which was sent by the bot.
}
