% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram_package.R
\docType{package}
\name{telegram}
\alias{telegram-package}
\alias{telegram}
\title{telegram.}
\description{
Simple R wrapper around the Telegram Bot API
(http://core.telegram.org/bots/api) to access Telegram's messaging
facilities with ease (eg you send messages, images, files from R to
your smartphone)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lbraglia/telegram}
  \item Report bugs at \url{https://github.com/lbraglia/telegram/issues}
}

}
\author{
\strong{Maintainer}: Luca Braglia \email{lbraglia@gmail.com}

}
