% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\docType{class}
\name{TGBot}
\alias{TGBot}
\title{TGBot}
\format{An \code{\link{R6Class}} generator object.}
\description{
Package main class (implementing the Telegram bot).
}
\section{API Methods}{
 \describe{
    \item{\code{\link{forwardMessage}}}{forward messages of any
    kind} \item{\code{\link{getFile}}}{get info about a file and
    download it} \item{\code{\link{getMe}}}{test your bot's auth
    token} \item{\code{\link{getUpdates}}}{receive incoming
    updates} \item{\code{\link{getUserProfilePhotos}}}{get a list
    of profile pictures for a user}
    \item{\code{\link{sendAudio}}}{send \code{mp3} files}
    \item{\code{\link{sendDocument}}}{send general files}
    \item{\code{\link{sendLocation}}}{send point on the map}
    \item{\code{\link{sendMessage}}}{send text messages}
    \item{\code{\link{sendPhoto}}}{send image files}
    \item{\code{\link{sendSticker}}}{send \code{.webp} stickers}
    \item{\code{\link{sendVideo}}}{send \code{mp4} videos}
    \item{\code{\link{sendVoice}}}{send ogg files encoded with
    OPUS} 
    \item{\code{\link{sendPoll}}}{send a telegram poll}
    \item{\code{\link{stopPoll}}}{stop a telegram poll and obtain results}
}
}

\examples{
\dontrun{

## Without proxy
bot <- TGBot$new(token = bot_token('RBot'))

## With a proxy ...
prx <- list('url' = '123.45.6.78',
            'port' = 8080,
            'username' = 'user',
            'password' = 'password')
bot <- TGBot$new(token = bot_token('RBot'), proxy = prx)
##  .. or 
bot <- TGBot$new(token = bot_token('RBot'))
bot$set_proxy(proxy = prx)

}
}
\references{
\href{http://core.telegram.org/bots}{Bots: An
    introduction for developers} and
    \href{http://core.telegram.org/bots/api}{Telegram Bot API}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TGBot$new()}}
\item \href{#method-set_token}{\code{TGBot$set_token()}}
\item \href{#method-set_proxy}{\code{TGBot$set_proxy()}}
\item \href{#method-set_default_chat_id}{\code{TGBot$set_default_chat_id()}}
\item \href{#method-req}{\code{TGBot$req()}}
\item \href{#method-last_request}{\code{TGBot$last_request()}}
\item \href{#method-print}{\code{TGBot$print()}}
\item \href{#method-forwardMessage}{\code{TGBot$forwardMessage()}}
\item \href{#method-getFile}{\code{TGBot$getFile()}}
\item \href{#method-getMe}{\code{TGBot$getMe()}}
\item \href{#method-getUpdates}{\code{TGBot$getUpdates()}}
\item \href{#method-getUserProfilePhotos}{\code{TGBot$getUserProfilePhotos()}}
\item \href{#method-sendAudio}{\code{TGBot$sendAudio()}}
\item \href{#method-sendChatAction}{\code{TGBot$sendChatAction()}}
\item \href{#method-sendDocument}{\code{TGBot$sendDocument()}}
\item \href{#method-sendLocation}{\code{TGBot$sendLocation()}}
\item \href{#method-sendMessage}{\code{TGBot$sendMessage()}}
\item \href{#method-sendPhoto}{\code{TGBot$sendPhoto()}}
\item \href{#method-sendSticker}{\code{TGBot$sendSticker()}}
\item \href{#method-sendVideo}{\code{TGBot$sendVideo()}}
\item \href{#method-sendVoice}{\code{TGBot$sendVoice()}}
\item \href{#method-sendDice}{\code{TGBot$sendDice()}}
\item \href{#method-sendPoll}{\code{TGBot$sendPoll()}}
\item \href{#method-stopPoll}{\code{TGBot$stopPoll()}}
\item \href{#method-setWebhook}{\code{TGBot$setWebhook()}}
\item \href{#method-clone}{\code{TGBot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$new(token, proxy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_token"></a>}}
\subsection{Method \code{set_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$set_token(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_proxy"></a>}}
\subsection{Method \code{set_proxy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$set_proxy(proxy)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_default_chat_id"></a>}}
\subsection{Method \code{set_default_chat_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$set_default_chat_id(chat_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-req"></a>}}
\subsection{Method \code{req()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$req(method = NULL, body = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-last_request"></a>}}
\subsection{Method \code{last_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$last_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forwardMessage"></a>}}
\subsection{Method \code{forwardMessage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$forwardMessage(
  from_chat_id = NULL,
  message_id = NULL,
  disable_notification = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFile"></a>}}
\subsection{Method \code{getFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$getFile(file_id, destfile = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMe"></a>}}
\subsection{Method \code{getMe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$getMe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUpdates"></a>}}
\subsection{Method \code{getUpdates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$getUpdates(offset = NULL, limit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUserProfilePhotos"></a>}}
\subsection{Method \code{getUserProfilePhotos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$getUserProfilePhotos(user_id = NULL, offset = NULL, limit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendAudio"></a>}}
\subsection{Method \code{sendAudio()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendAudio(
  audio = NULL,
  caption = NULL,
  duration = NULL,
  performer = NULL,
  title = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendChatAction"></a>}}
\subsection{Method \code{sendChatAction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendChatAction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendDocument"></a>}}
\subsection{Method \code{sendDocument()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendDocument(
  document = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendLocation"></a>}}
\subsection{Method \code{sendLocation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendLocation(
  latitude = NULL,
  longitude = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendMessage"></a>}}
\subsection{Method \code{sendMessage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendMessage(
  text = NULL,
  parse_mode = NULL,
  disable_web_page_preview = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendPhoto"></a>}}
\subsection{Method \code{sendPhoto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendPhoto(
  photo = NULL,
  caption = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendSticker"></a>}}
\subsection{Method \code{sendSticker()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendSticker(
  sticker = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendVideo"></a>}}
\subsection{Method \code{sendVideo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendVideo(
  video = NULL,
  duration = NULL,
  caption = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendVoice"></a>}}
\subsection{Method \code{sendVoice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendVoice(
  voice = NULL,
  duration = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendDice"></a>}}
\subsection{Method \code{sendDice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendDice(disable_notification = NULL, chat_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendPoll"></a>}}
\subsection{Method \code{sendPoll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendPoll(
  question = NULL,
  options = NULL,
  is_anonymous = TRUE,
  type = c("regular", "quiz"),
  allows_multiple_answers = FALSE,
  correct_option_id = NULL,
  is_closed = FALSE,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stopPoll"></a>}}
\subsection{Method \code{stopPoll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$stopPoll(message_id = NULL, chat_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setWebhook"></a>}}
\subsection{Method \code{setWebhook()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$setWebhook()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
