% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_telegram_envvar.R
\name{bot_token}
\alias{bot_token}
\title{bot_token}
\usage{
bot_token(x)
}
\arguments{
\item{x}{character of length 1 with the name of the bot; if
\code{NULL} a menu to choose between bot is displayed and the
proper token returned}
}
\description{
Obtain token from system variables (in \code{Renviron}) set
according to package naming conventions, that is
\code{R_TELEGRAM_BOT_} where \code{X} is bot's name (first question
answered to the botfather).
}
\examples{
\dontrun{ bot_token('RBot') }
}

