% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{getUpdates}
\alias{getUpdates}
\title{getUpdates}
\usage{
getUpdates(offset = NULL, limit = NULL)
}
\arguments{
\item{offset}{Identifier of the first update to be returned
returned.}

\item{limit}{Limits the number of updates to be retrieved. Values
between 1-100 are accepted. Defaults to 100}
}
\description{
Receive incoming updates
}

