% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendAudio}
\alias{sendAudio}
\title{sendAudio}
\usage{
sendAudio(audio = NULL, duration = NULL, performer = NULL, title = NULL,
  reply_to_message_id = NULL, chat_id = NULL)
}
\arguments{
\item{audio}{path to audio file to send (required)}

\item{duration}{duration of the audio in seconds}

\item{performer}{performer}

\item{title}{track name}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Send \code{mp3} files
}

