% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendChatAction}
\alias{sendChatAction}
\title{sendChatAction}
\usage{
sendChatAction(action = NULL, chat_id = NULL)
}
\arguments{
\item{action}{Type of action to broadcast. Choose one, depending on
what the user is about to receive:
\itemize{
 \item{'\emph{typing}'}{ for text messages}
 \item{'\emph{upload_photo}'}{ for photos}
 \item{'\emph{upload_video}'}{ for videos}
 \item{'\emph{record_video}'}{ for video recording}
 \item{'\emph{upload_audio}'}{ for audio files}
 \item{'\emph{record_audio}'}{ for audio file recording}
 \item{'\emph{upload_document}'}{ for general files}
 \item{'\emph{find_location}'}{ for location data}
 \item{'\emph{upload_video_note}'}{ for video notes}
 \item{'\emph{record_video_note}'}{ for video note recording}
}}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Use this method when you need to tell the user that something is
happening on the bot's side. The status is set for 5 seconds or
less (when a message arrives from your bot, Telegram clients clear
its typing status).
}
