% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendChatAction}
\alias{sendChatAction}
\title{sendChatAction}
\usage{
sendChatAction(action = NULL, chat_id = NULL)
}
\arguments{
\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Use this method when you need to tell the user that something is
happening on the bot's side. The status is set for 5 seconds or
less (when a message arrives from your bot, Telegram clients clear
its typing status).
}
