% Generated by ebeneditos: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendChatAction}
\alias{sendChatAction}
\title{sendChatAction}
\usage{
sendChatAction(action = NULL, chat_id = NULL)
}
\arguments{
\item{action}{Type of action to broadcast (required). Choose one, depending on
what the user is about to receive: *typing* for text messages,
*upload_photo* for photos, *record_video* or *upload_video* for videos,
*record_audio* or *upload_audio* for audio files, *upload_document* for
general files, *find_location* for location data, *record_video_note*
or *upload_video_note* for video notes.}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Send text messages
}

