% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendDice}
\alias{sendDice}
\title{sendDice}
\usage{
sendDice(chat_id = NULL)
}
\arguments{
\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
sendDice
}
