% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendLocation}
\alias{sendLocation}
\title{sendLocation}
\usage{
sendLocation(
  latitude = NULL,
  longitude = NULL,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)
}
\arguments{
\item{latitude}{Latitude of location (required)}

\item{longitude}{Longitude of location (required)}

\item{disable_notification}{Sends the message silently. Users will
receive a notification with no sound}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Send point on the map
}
