% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\name{sendPoll}
\alias{sendPoll}
\title{sendPoll}
\usage{
sendPoll(
  question = NULL,
  options = NULL,
  is_anonymous = TRUE,
  type = c("regular", "quiz"),
  allows_multiple_answers = FALSE,
  correct_option_id = NULL,
  is_closed = FALSE,
  disable_notification = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)
}
\arguments{
\item{question}{Poll question, 1-255 characters}

\item{options}{vector of possible answers: 2-10 strings 1-100
characters each}

\item{is_anonymous}{if TRUE (default) the poll needs to be
anonymous}

\item{type}{poll type, 'quiz' or 'regular', defaults to 'regular'}

\item{allows_multiple_answers}{if TRUE the poll allows multiple
answers (default to FALSE and ignored for polls in quiz mode)}

\item{correct_option_id}{0-based identifier of the correct answer
option, required for polls in quiz mode (eg 0 set the first
answer as correct)}

\item{is_closed}{if TRUE the poll needs to be immediately
closed. This can be useful for poll preview. (default to FALSE)}

\item{disable_notification}{Sends the message silently. Users will
receive a notification with no sound}

\item{reply_to_message_id}{If the message is a reply, ID of the
original message}

\item{chat_id}{Unique identifier for the target chat or username of
the target channel (required)}
}
\description{
Use this method to send a native poll.
}
