% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram_package.R
\docType{package}
\name{telegram}
\alias{telegram}
\alias{telegram-package}
\title{telegram.}
\description{
R wrapper around the Telegram Bot API
(http://core.telegram.org/bots/api) to access Telegram's messaging
facilities with ease (eg you send messages, images, files from R to
your smartphone)
}

