% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\docType{class}
\name{TGBot}
\alias{TGBot}
\title{TGBot}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
Package main class (implementing the Telegram bot).
}
\section{API Methods}{
 \describe{
    \item{\code{\link{forwardMessage}}}{forward messages of any
    kind} \item{\code{\link{getFile}}}{get info about a file and
    download it} \item{\code{\link{getMe}}}{test your bot's auth
    token} \item{\code{\link{getUpdates}}}{receive incoming
    updates} \item{\code{\link{getUserProfilePhotos}}}{get a list
    of profile pictures for a user}
    \item{\code{\link{sendAudio}}}{send \code{mp3} files}
    \item{\code{\link{sendDocument}}}{send general files}
    \item{\code{\link{sendLocation}}}{send point on the map}
    \item{\code{\link{sendMessage}}}{send text messages}
    \item{\code{\link{sendPhoto}}}{send image files}
    \item{\code{\link{sendSticker}}}{send \code{.webp} stickers}
    \item{\code{\link{sendVideo}}}{send \code{mp4} videos}
    \item{\code{\link{sendVoice}}}{send ogg files encoded with
    OPUS} }
}

\examples{
\dontrun{
bot <- TGBot$new(token = bot_token('RBot'))
}
}
\references{
\href{http://core.telegram.org/bots}{Bots: An
    introduction for developers} and
    \href{http://core.telegram.org/bots/api}{Telegram Bot API}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TGBot$new()}}
\item \href{#method-set_token}{\code{TGBot$set_token()}}
\item \href{#method-set_default_chat_id}{\code{TGBot$set_default_chat_id()}}
\item \href{#method-last_request}{\code{TGBot$last_request()}}
\item \href{#method-print}{\code{TGBot$print()}}
\item \href{#method-forwardMessage}{\code{TGBot$forwardMessage()}}
\item \href{#method-getFile}{\code{TGBot$getFile()}}
\item \href{#method-getMe}{\code{TGBot$getMe()}}
\item \href{#method-getUpdates}{\code{TGBot$getUpdates()}}
\item \href{#method-getUserProfilePhotos}{\code{TGBot$getUserProfilePhotos()}}
\item \href{#method-sendAudio}{\code{TGBot$sendAudio()}}
\item \href{#method-sendChatAction}{\code{TGBot$sendChatAction()}}
\item \href{#method-sendDocument}{\code{TGBot$sendDocument()}}
\item \href{#method-sendLocation}{\code{TGBot$sendLocation()}}
\item \href{#method-sendMessage}{\code{TGBot$sendMessage()}}
\item \href{#method-sendPhoto}{\code{TGBot$sendPhoto()}}
\item \href{#method-sendSticker}{\code{TGBot$sendSticker()}}
\item \href{#method-sendVideo}{\code{TGBot$sendVideo()}}
\item \href{#method-sendVoice}{\code{TGBot$sendVoice()}}
\item \href{#method-setWebhook}{\code{TGBot$setWebhook()}}
\item \href{#method-clone}{\code{TGBot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$new(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_token"></a>}}
\if{latex}{\out{\hypertarget{method-set_token}{}}}
\subsection{Method \code{set_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$set_token(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_default_chat_id"></a>}}
\if{latex}{\out{\hypertarget{method-set_default_chat_id}{}}}
\subsection{Method \code{set_default_chat_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$set_default_chat_id(chat_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-last_request"></a>}}
\if{latex}{\out{\hypertarget{method-last_request}{}}}
\subsection{Method \code{last_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$last_request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forwardMessage"></a>}}
\if{latex}{\out{\hypertarget{method-forwardMessage}{}}}
\subsection{Method \code{forwardMessage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$forwardMessage(from_chat_id = NULL, message_id = NULL, chat_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFile"></a>}}
\if{latex}{\out{\hypertarget{method-getFile}{}}}
\subsection{Method \code{getFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$getFile(file_id, destfile = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMe"></a>}}
\if{latex}{\out{\hypertarget{method-getMe}{}}}
\subsection{Method \code{getMe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$getMe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUpdates"></a>}}
\if{latex}{\out{\hypertarget{method-getUpdates}{}}}
\subsection{Method \code{getUpdates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$getUpdates(offset = NULL, limit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUserProfilePhotos"></a>}}
\if{latex}{\out{\hypertarget{method-getUserProfilePhotos}{}}}
\subsection{Method \code{getUserProfilePhotos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$getUserProfilePhotos(user_id = NULL, offset = NULL, limit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendAudio"></a>}}
\if{latex}{\out{\hypertarget{method-sendAudio}{}}}
\subsection{Method \code{sendAudio()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendAudio(
  audio = NULL,
  caption = NULL,
  duration = NULL,
  performer = NULL,
  title = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendChatAction"></a>}}
\if{latex}{\out{\hypertarget{method-sendChatAction}{}}}
\subsection{Method \code{sendChatAction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendChatAction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendDocument"></a>}}
\if{latex}{\out{\hypertarget{method-sendDocument}{}}}
\subsection{Method \code{sendDocument()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendDocument(document = NULL, reply_to_message_id = NULL, chat_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendLocation"></a>}}
\if{latex}{\out{\hypertarget{method-sendLocation}{}}}
\subsection{Method \code{sendLocation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendLocation(
  latitude = NULL,
  longitude = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendMessage"></a>}}
\if{latex}{\out{\hypertarget{method-sendMessage}{}}}
\subsection{Method \code{sendMessage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendMessage(
  text = NULL,
  parse_mode = NULL,
  disable_web_page_preview = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendPhoto"></a>}}
\if{latex}{\out{\hypertarget{method-sendPhoto}{}}}
\subsection{Method \code{sendPhoto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendPhoto(
  photo = NULL,
  caption = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendSticker"></a>}}
\if{latex}{\out{\hypertarget{method-sendSticker}{}}}
\subsection{Method \code{sendSticker()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendSticker(sticker = NULL, reply_to_message_id = NULL, chat_id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendVideo"></a>}}
\if{latex}{\out{\hypertarget{method-sendVideo}{}}}
\subsection{Method \code{sendVideo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendVideo(
  video = NULL,
  duration = NULL,
  caption = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendVoice"></a>}}
\if{latex}{\out{\hypertarget{method-sendVoice}{}}}
\subsection{Method \code{sendVoice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$sendVoice(
  voice = NULL,
  duration = NULL,
  reply_to_message_id = NULL,
  chat_id = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setWebhook"></a>}}
\if{latex}{\out{\hypertarget{method-setWebhook}{}}}
\subsection{Method \code{setWebhook()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$setWebhook()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TGBot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
