% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\docType{class}
\name{TGBot}
\alias{TGBot}
\title{TGBot}
\format{An \code{\link{R6Class}} generator object.}
\usage{
TGBot
}
\description{
Package main class (implementing the Telegram bot).
}
\section{API Methods}{

\describe{
  \item{\code{getMe}}{tests your bot's auth token}
  \item{\code{sendMessage}}{Send text messages}
  \item{\code{forwardMessage}}{Forward messages of any kind}
  \item{\code{sendPhoto}}{Send image files.}
  \item{\code{sendAudio}}{Send \code{mp3} files}
  \item{\code{sendDocument}}{Send general files}
  \item{\code{sendSticker}}{Send \code{.webp} stickers}
  \item{\code{sendVideo}}{Send \code{mp4} videos}
  \item{\code{sendVoice}}{Send ogg files encoded with OPUS}
  \item{\code{sendLocation}}{Send point on the map}
  \item{\code{getUserProfilePhotos}}{Get a list of profile pictures for a user}
  \item{\code{getFile}}{Get basic info about a file and prepare it for downloading}
}
}
\examples{
\dontrun{
## Talk to the botfather and create a new bot following the steps
## ...
## After you've done, put the returned token in the following command
## to handle the bot
bot <- TGBot$new(token = get_tgbot_token('RBot'))

## Now check bot connection it should print some of your bot's data
bot$getMe()

## Now find and say something to your bot on the phone to start a chat
## (and obtain a chat id).
## ...
## Here, check what you have inserted
bot$getUpdates()

## You're interested in the message.chat.id variable: in order to set a
## default chat_id for the following commands (to ease typing)
bot$set_default_chat_id(123456789)

## Send some messages..
bot$sendMessage('This is text')
## Markdown support for messages
md1 <- "*bold* _italic_ [r-project](http://r-project.org) "
md2 <- " try `x <- rnorm(100)` at the console ..."
md3 <- "
you can have
```
x <- runif(100)
mean(x)
```
too
"
bot$sendMessage(md1, parse_mode = 'markdown')
bot$sendMessage(md2, parse_mode = 'markdown')
bot$sendMessage(md3, parse_mode = 'markdown')

## Send a image/photo
png('test.png')
plot(rnorm(100))
dev.off()
bot$sendPhoto('test.png', caption = 'This is my awesome graph')

## Send a document (can be any file)
help(TGBot, help_type = 'pdf')
bot$sendDocument('TGBot.pdf')

## Forward a message
bot$forwardMessage(from_chat_id = 123456,
                   chat_id = 123456,
                   message_id = 35)

## Send a location
bot$sendLocation(44.699, 10.6297)

## Send a sticker
bot$sendSticker(system.file('r_logo.webp', package = 'telegram'))

## Send a video
library(animation)
saveVideo({
    set.seed(1)
    nmax <- 10
    ani.options(interval = 0.4, nmax = nmax)
    x <- c()
    for (i in 1:nmax){
        x <- c(x, rnorm(1))
        plot(cumsum(x), lty = 2, xlim = c(1, nmax), ylim = c(-5,5))
        abline(h = 0, col = 'red')
    }
}, video.name = 'animation.mp4')
bot$sendVideo('animation.mp4')

## Send mp3 audio files
bot$sendAudio(system.file('audio_test.mp3', package = 'telegram'),
              performer = 'espeak (http://espeak.sf.net)')

## Send voice (opus encoded .ogg files)
bot$sendVoice(system.file('voice_test.ogg', package = 'telegram'))
}

}
\references{
\href{http://core.telegram.org/bots}{Bots: An introduction for developers} and \href{http://core.telegram.org/bots/api}{Telegram Bot API}
}
\keyword{datasets}

