% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGBot.R
\docType{class}
\name{TGBot}
\alias{TGBot}
\title{TGBot}
\format{An \code{\link{R6Class}} generator object.}
\usage{
TGBot
}
\description{
Package main class (implementing the Telegram bot).
}
\section{API Methods}{

\describe{
  \item{\code{getMe}}{tests your bot's auth token}
  \item{\code{sendMessage}}{Send text messages}
  \item{\code{forwardMessage}}{Forward messages of any kind}
  \item{\code{sendPhoto}}{Send image files.}
  \item{\code{sendAudio}}{Send \code{mp3} files}
  \item{\code{sendDocument}}{Send general files}
  \item{\code{sendSticker}}{Send \code{.webp} stickers}
  \item{\code{sendVideo}}{Send \code{mp4} videos}
  \item{\code{sendVoice}}{Send ogg files encoded with OPUS}
  \item{\code{sendLocation}}{Send point on the map}
  \item{\code{getUserProfilePhotos}}{Get a list of profile pictures for a user}
  \item{\code{getFile}}{Get basic info about a file and prepare it for downloading}
}
}
\examples{
\dontrun{
bot <- TGBot$new(token = bot_token('RBot'))
}
}
\references{
\href{http://core.telegram.org/bots}{Bots: An introduction for developers} and \href{http://core.telegram.org/bots/api}{Telegram Bot API}
}
\keyword{datasets}

